requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
colorama==0.4.3
Django==3.0.8
django-fontawesome==1.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
Pillow==7.2.0
pycodestyle==2.6.0
pylint==2.5.3
PyMySQL==0.10.1
python-decouple==3.4
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
